make gap histogram buckets appropriate for milliseconds
The taskrun gap histogram metric still had the bucket specification that
was copied and pasted from our existing histograms measured in seconds.
Multipled by 1000 as needed to shift the buckets appropriately.

Also added a 'status' label with 'succeeded' and 'failed' values to correlate
with a similar label available on the upstream pipelineruns duration metric,
in case we see patterns where gaps vary based on failures in the underlying
containers.