PLNSRVCE-1585: flag throttled pipelineruns; filter overhead alert metric when throttled (#52)
* go tidy needed

* Tag PipelinRuns that have been throttled; filter throttled PipelineRun from overhead metrics

So recent bugs and inefficiencies in RHTAP have lead to stampedes of PipelineRuns that
violated quota to the extreme.  This impacts pipeline controller performance not just because of shared cpu/mem, but becase Tekton requeues Pod creation error events, backing up reconcilation, and leading to greater overhead.  As we already have explicit metrics for quota and node violations, we will shield the execution overhead metric, which aims to see if pipeline controller is overly falling behind when PipelineRuns are evenly spread across the cluster.

Also some tweaks to some short term tracing were made.