add metric to determine if the core controller is not attempting to create underlying pods
While the current execution overhead metric can indicate that the core Tekton controllers
are struggling, it is not a definitive declaration that the controller is deadlocked.

Ultimately, Tekton is deadlocked if the TaskRun is not in terminal state, but it has not even
attempted to create the underlying Pod.

End of the day, Tekton core controller needs to create Pods from the higher level abstraction
API to run the user's Pipelines.

We add the metric to capture this with this PR.

rh-pre-commit.version: 2.2.0
rh-pre-commit.check-secrets: ENABLED